added controllers, services, intefaces of firma, preventivo, user

+++ backend/FlliBrutti.Backend.Application/Models/UserDTO.cs
@@ -14,7 +14,7 @@ public class UserDTO : PersonDTO
     {
         return new User
         {
-            Type = Type,
+            Type = (int)Type,
             Email = Email,
             Password = Password
         };
